Refactor: User,Auth 경로 리팩토링 Auth에 대한 테스트 코드 경로가 맞지 않아 리팩토링 했습니다. User ErrorCode도 global에서 user.error 경로로 리팩토링 했습니다.
